Multiple active gateway enhancement proposal (#69)
* Multiple active gateway enhancement proposal

Submariner currently only allows a single gateway to be active at
any one time in a given cluster.  This enhancement proposes that
there be an option to allow more than one gateway to be active in
a given cluster to enhance both performance and resiliency. This
feature is sometimes referred to as “active-active gateways”;
however, the goal is to allow two or more actove gateways to be used.

Signed-off-by: Andre Fredette <afredette@redhat.com>
Co-authored-by: Andrew Stoycos <astoycos@redhat.com>
Co-authored-by: Billy McFall <22157057+Billy99@users.noreply.github.com>