Run dependent issues check only on open/edit
Instead of running it daily, which still seems to be too often.

This repo has a lot of activity relative to the number of commits.
We think this causes the dependent issue check to run on the same SHA too
many times and start failing with:

> SHA and context has reached the maximum number of statuses

This commit will also make a new SHA, which should fix the issue.

Signed-off-by: Daniel Farrell <dfarrell@redhat.com>