feat(subctl): add `upgrade` sub-command to automate Submariner upgrade process
This commit adds a new `upgrade` sub-command to `subctl` to automate the process of upgrading Submariner. The `upgrade` command takes a `to` flag to specify the version of Submariner to upgrade to. Once the command is run, `subctl` will upgrade itself to the specified version and then re-run `deploy-broker` and `join` sub-commands on all contexts where Broker and Submariner are installed respectively.

This feature addresses the issue #183.

Signed-Off-By: Janki Chhatbar <jchhatba@redhat.com>